services/lambda_autoscaling/requirements.txt
boto3>=1.26.0
python-dotenv>=1.0.0
redis>=5.0.0